requirements.txt
# Core ML / NLP
torch>=1.10.0
transformers>=4.30.0
datasets>=2.0.0

# Data
numpy>=1.21.0
pandas>=1.5.0
openpyxl>=3.0.0          # 读取 .xlsx 文件（Excel gold data）

# Metrics
scikit-learn>=1.0.0

# Progress bar
tqdm>=4.60.0

# Config
pyyaml>=6.0               # 读取 configs/tasks.yaml
